services/ingestor/requirements.txt
aiohttp
kafka-python-ng
asyncio